requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.8
django-phonenumber-field==5.2.0
gunicorn==20.1.0
pdfkit==0.6.1
phonenumbers==8.12.36
Pillow==8.4.0
python-dotenv==0.19.1
pytz==2021.3
qrcode==7.3.1
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
